+++ coeus-autograd/src/ops/mod.rs
@@ -31,10 +31,10 @@ pub use arithmetic::VarScalarExt;
 pub use linalg::{matmul, sparse_matmul, sparse_matmul_coo, transpose_2d};
 pub use nn::{
     avg_pool2d, avg_pool3d, batchnorm1d, batchnorm2d, batchnorm3d, binary_cross_entropy, conv1d,
-    conv2d, conv3d, conv_transpose1d, conv_transpose2d, cosine_embedding_loss, cross_entropy_loss,
-    dropout, huber_loss, kl_divergence, layernorm, log_softmax, margin_ranking_loss, max_pool2d,
-    max_pool3d, nll_loss, rmsnorm, sdp_attention, softmax, AttentionMask, BatchNormArgs,
-    CausalMask, NullMask,
+    conv2d, conv3d, conv_transpose1d, conv_transpose2d, conv_transpose3d, cosine_embedding_loss,
+    cross_entropy_loss, dropout, huber_loss, kl_divergence, layernorm, log_softmax,
+    margin_ranking_loss, max_pool2d, max_pool3d, nll_loss, rmsnorm, sdp_attention, softmax,
+    AttentionMask, BatchNormArgs, CausalMask, NullMask,
 };
 
 pub use embedding::{embedding, embedding_with_padding_idx};