+++ coeus-nn/src/conv/mod.rs
@@ -4,6 +4,7 @@ pub mod dim;
 mod conv_nd;
 mod conv_transpose1d;
 mod conv_transpose2d;
+mod conv_transpose3d;
 
 pub use conv_nd::{Conv, ConvParams};
 pub use dim::{ConvDim, Dim1D, Dim2D, Dim3D};
@@ -17,3 +18,4 @@ pub type Conv3d<T, B = coeus_core::MoiraiBackend> = Conv<T, B, Dim3D>;
 
 pub use conv_transpose1d::ConvTranspose1d;
 pub use conv_transpose2d::ConvTranspose2d;
+pub use conv_transpose3d::ConvTranspose3d;